#  Upgrade Elasticsearch version to 5.6.10 and Accumulo to 1.9.2