--- a/accumulo/src/main/java/org/vertexium/accumulo/mapreduce/AccumuloElementOutputFormat.java
+++ b/accumulo/src/main/java/org/vertexium/accumulo/mapreduce/AccumuloElementOutputFormat.java
@@ -15,7 +15,7 @@ public class AccumuloElementOutputFormat extends OutputFormat<Text, Mutation> {
 
     public static void setOutputInfo(Job job, String instanceName, String zooKeepers, String principal, AuthenticationToken token) throws AccumuloSecurityException {
         AccumuloOutputFormat.setConnectorInfo(job, principal, token);
-        ClientConfiguration clientConfig = new ClientConfiguration()
+        ClientConfiguration clientConfig = ClientConfiguration.create()
                 .withInstance(instanceName)
                 .withZkHosts(zooKeepers);
         AccumuloOutputFormat.setZooKeeperInstance(job, clientConfig);