--- a/elasticsearch5-plugin/src/main/assemblies/plugin-descriptor.properties
+++ b/elasticsearch5-plugin/src/main/assemblies/plugin-descriptor.properties
@@ -2,5 +2,5 @@ description=Vertexium
 version=master
 classname=org.vertexium.elasticsearch5.plugin.VertexiumElasticsearchPlugin
 java.version=1.8
-elasticsearch.version=5.5.3
+elasticsearch.version=5.6.10
 name=vertexium