--- a/elasticsearch5-plugin/src/main/java/org/vertexium/elasticsearch5/plugin/VertexiumQueryStringQueryBuilder.java
+++ b/elasticsearch5-plugin/src/main/java/org/vertexium/elasticsearch5/plugin/VertexiumQueryStringQueryBuilder.java
@@ -103,7 +103,7 @@ public class VertexiumQueryStringQueryBuilder extends QueryStringQueryBuilder {
             }
 
             return FieldNameToVisibilityMap.createFromVertexiumMetadata(results);
-        } catch (IOException ex) {
+        } catch (Exception ex) {
             throw new RuntimeException("Could not get mappings", ex);
         }
     }