Wanted Accumulo 1.9 to get faster performance https://accumulo.apache.org/release/accumulo-1.9.0/#performance-improvements

According to `GraphTestBase#benchmark` running Accumulo

Before...

```
2018-08-01 11:39:21.913/EDT INFO  [vertexium.test.GraphTestBase] add vertices in 0.934s
2018-08-01 11:39:22.784/EDT INFO  [vertexium.test.GraphTestBase] add edges in 0.871s
2018-08-01 11:39:43.282/EDT INFO  [vertexium.test.GraphTestBase] find vertices by id in 20.498s
```

After...

```
2018-08-01 10:25:33.749/EDT INFO  [vertexium.test.GraphTestBase] add vertices in 1.074s
2018-08-01 10:25:34.841/EDT INFO  [vertexium.test.GraphTestBase] add edges in 1.092s
2018-08-01 10:25:50.887/EDT INFO  [vertexium.test.GraphTestBase] find vertices by id in 16.045s
```

So it looks like writes were slightly slower but reads were quite a bit faster.